+++ src/components/toolbar/GraphButton/Button.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import GraphIcon from '@src/utils/svg/graph.svg';
+import * as StyledComponent from './style';
+
+interface Props {
+  onClick: () => void;
+}
+const GraphButton = ({ onClick }: Props) => {
+  return (
+    <StyledComponent.ButtonContainer onClick={onClick}>
+      <GraphIcon />
+    </StyledComponent.ButtonContainer>
+  );
+};
+
+export default GraphButton;
