+++ src/components/toolbar/GraphButton/style.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+interface Props {
+  show: boolean;
+}
+export const ButtonContainer = styled.div`
+  background-color: transparent;
+  cursor: pointer;
+  svg {
+    background-color: transparent;
+  }
+`;
